Ensure odd primes aren't defined when `odd-primes` is disabled
